# api/v2/manufacturer/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FDO Manufacturer Server API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Go FDO Server API Support",
      "url": "https://github.com/fido-device-onboard/go-fdo-server/issues"
    }
  },
  "paths": {
...
    "/api/v1/vouchers": {
      "get": {
        "operationId": "ListVouchers",
        "summary": "Query ownership vouchers (Deprecated)",
        "description": "Returns a list of ownership vouchers matching optional filter criteria.\nResults are ordered by updated_at timestamp in descending order.\nThe CBOR field is omitted from results (use GET /vouchers/{guid} to retrieve full voucher).\n",
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "description": "Filter by device GUID (hex-encoded)",
            "schema": {
              "type": "string",
              "example": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c",
              "pattern": "^[0-9a-fA-F]{32}$"
            }
          },
          {
            "name": "device_info",
            "in": "query",
            "description": "Filter by device info string",
            "schema": {
              "type": "string",
              "example": "MyDevice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved voucher list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoucherMetadata"
                  }
                },
                "examples": {
                  "emptyList": {
                    "summary": "No vouchers found",
                    "value": []
                  },
                  "voucherList": {
                    "summary": "List of vouchers",
                    "value": [
                      {
                        "guid": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c",
                        "device_info": "Device 1",
                        "created_at": "2024-01-15T10:30:00Z",
                        "updated_at": "2024-01-15T10:30:00Z"
                      },
                      {
                        "guid": "a1b2c3d4e5f6789012345678abcdef01",
                        "device_info": "Device 2",
                        "created_at": "2024-01-14T09:20:00Z",
                        "updated_at": "2024-01-14T09:20:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "example": "Invalid GUID: xyz"
              }
            }
          },
          "500": {
            "description": "Internal server error occurred while querying vouchers",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                }
              }
            }
          }
        },
        "tags": [
          "vouchers"
        ]
      },
      "post": {
        "operationId": "InsertVoucher",
        "summary": "Insert ownership voucher(s) (Deprecated)",
        "description": "Inserts one or more ownership vouchers into the database.\nThe request body should contain PEM-encoded ownership voucher(s).\nEach voucher is verified for:\n- Ownership (voucher owner key matches server's configured keys)\n- Cryptographic integrity (signature chain, certificate chains, etc.)\n- Protocol version and header field validity\nDuplicate vouchers (same GUID with same CBOR data) are silently skipped.\nExisting vouchers (same GUID with different CBOR data) are not overwritten and are silently skipped.\nThe operation succeeds (200 OK) even if all vouchers were duplicates/skipped.\n",
        "requestBody": {
          "description": "PEM-encoded ownership voucher(s)",
          "required": true,
          "content": {
            "application/x-pem-file": {
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "example": "-----BEGIN OWNERSHIP VOUCHER-----\n[base64-encoded CBOR data]\n-----END OWNERSHIP VOUCHER-----\n"
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voucher(s) successfully inserted or were duplicates",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "400": {
            "description": "Invalid voucher data or verification failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "examples": {
                  "decodeCborError": {
                    "summary": "Unable to decode CBOR",
                    "value": "Unable to decode cbor"
                  },
                  "invalidVoucher": {
                    "summary": "Voucher verification failed",
                    "value": "Invalid ownership voucher"
                  },
                  "decodePemError": {
                    "summary": "Unable to decode PEM",
                    "value": "Unable to decode PEM content"
                  }
                }
              }
            }
          },
          "413": {
            "description": "Request payload too large (exceeds maximum allowed size).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "example": "http: request body too large"
              }
            }
          },
          "500": {
            "description": "Internal server error occurred while inserting voucher",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "examples": {
                  "readBodyError": {
                    "summary": "Error reading request body",
                    "value": "Failure to read the request body"
                  },
                  "insertError": {
                    "summary": "Error inserting into database",
                    "value": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "vouchers"
        ]
      }
    },
...
    "/api/v2/vouchers/{guid}/extend": {
      "post": {
        "operationId": "ExtendOwnershipVoucher",
        "summary": "Extend an ownership voucher",
        "description": "Extends an ownership voucher by adding a new ownership entry. This operation\nis used to transfer ownership of a device from one owner to another.\n\nThe request must include the new owner's public key and will be signed\nwith the current owner's private key.\n",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "description": "The device GUID (32 hexadecimal characters).",
            "required": true,
            "schema": {
              "type": "string",
              "example": "d21d841a3f54f4e89a60ed9b9779e9e8",
              "pattern": "^[a-f0-9]{32}$"
            }
          }
        ],
        "requestBody": {
          "description": "New owner's public key or certificate in PEM format.",
          "required": true,
          "content": {
            "application/x-pem-file": {
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "examples": {
                "extendWithPublicKey": {
                  "summary": "Extend voucher with PEM encoded public key",
                  "value": "-----BEGIN PUBLIC KEY-----\nMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEaXhXuQUiGqKv5VqJ3VZqDz7XJ8v5\nKxMfLgH5Z3YqN8PzGqKJ5vZ3YqN8PzGqKJ5vZ3YqN8PzGqKJ5vZ3YqN8Pw==\n-----END PUBLIC KEY-----\n"
                },
                "extendWithCertificate": {
                  "summary": "Extend voucher with PEM encoded certificate",
                  "value": "-----BEGIN CERTIFICATE-----\nMIICljCCAX4CCQDqJ0HQqh0W5jANBgkqhkiG9w0BAQsFADANMQswCQYDVQQGEwJV\nUzAeFw0yNTEwMjcxNDAwMDBaFw0yNjEwMjcxNDAwMDBaMA0xCzAJBgNVBAYTAlVT\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAx2q5VqJ3VZqDz7XJ8v5K\nxMfLgH5Z3YqN8PzGqKJ5vZ3YqN8PzGqKJ5vZ3YqN8PzGqKJ5vZ3YqN8PwIDAQAB\nMA0GCSqGSIb3DQEBCwUAA4IBAQBqKJ5vZ3YqN8PzGqKJ5vZ3YqN8PzGqKJ5vZ3Yq\nN8PzGqKJ5vZ3YqN8Pw==\n-----END CERTIFICATE-----\n"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ownership voucher extended successfully.",
            "content": {
              "application/x-pem-file": {
                "schema": {
                  "description": "The extended voucher in PEM format.",
                  "type": "string",
                  "format": "binary"
                },
                "examples": {
                  "extendedVoucher": {
                    "summary": "Extended voucher in PEM format",
                    "value": "-----BEGIN OWNERSHIP VOUCHER-----\noWRndWlkUNIdhBo/VPTompDtm5d56egEaGVhZGVyWEKhBRkBAQZYPKEBGQEB\nAligGdpx5PoC1AO8aTU8kz1mJ9n1KZQ2b6xLvE/sWmPRBIVcEU3hhcmUgZGV2\naWNlIGluZm9ybWF0aW9uBWJtZnJLZXlYa6EBAmhta2V5dmFsdWUGaGNlcnRD\naGFpbkhhc2ihaEFsZ29yaXRobUhTSEEtMzg0aVZhbHVlWDCwkz1mJ9n1KZQ2\nb6xLvE/sWmPRBIVcEU3hhcmUgZGV2aWNlIGluZm9ybWF0aW9uBWJtZnJLZXlY\na6EBAmhta2V5dmFsdWUGaGNlcnRDaGFpbkhhc2ihaEFsZ29yaXRobUhTSEEt\nMzg0aVZhbHVlWDCwkz1mJ9n1KZQ2b6xLvE/sWmPRBIVcEU3hhcmUgZGV2aWNl\nIGluZm9ybWF0aW9uBWJtZnJLZXk=\n-----END OWNERSHIP VOUCHER-----\n"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "voucher"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VoucherMetadata": {
        "description": "Ownership voucher metadata (excludes CBOR data)",
        "type": "object",
        "properties": {
          "guid": {
            "description": "Device GUID (hex-encoded)",
            "type": "string",
            "example": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c"
          },
          "device_info": {
            "description": "Device information string",
            "type": "string",
            "example": "Intel NUC Model XYZ"
          },
          "created_at": {
            "description": "Timestamp when voucher was first inserted",
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "updated_at": {
            "description": "Timestamp when voucher was last updated",
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": [
          "guid",
          "device_info",
          "created_at",
          "updated_at"
        ]
      },
      "ErrorString": {
        "description": "A human-readable error message.",
        "type": "string"
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Bad Request - The request was malformed or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "badRequest": {
                "summary": "Example of a bad request error",
                "value": {
                  "message": "Invalid Request."
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Authentication credentials missing, invalid, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "unauthorized": {
                "summary": "Example of an unauthorized error",
                "value": {
                  "message": "Authentication token is invalid or has expired."
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - The authenticated user does not have permission to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "forbidden": {
                "summary": "Example of a forbidden error",
                "value": {
                  "message": "User does not have the right access level."
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}